fixbsp using commandline instead of switching to GD UI

+++ fixbsp.sh
@@ -1,12 +1,21 @@
 #!/bin/bash
 set -e
 
+
+# Find Godot app path
+godotapp=$(mdfind "kMDItemKind == 'Application' && kMDItemFSName == 'Godot*' && kMDItemContentTypeTree == 'public.executable'" | head -n 1)
+if [ -z "$godotapp" ]; then
+    echo "Error: Godot application not found."
+    exit 1
+fi
+godot="$godotapp/Contents/MacOS/Godot"
+echo "Found Godot app at: $godot"
+
 echo "You need to close GoDot before running this script. I will kill it with force for you if it won't die peacefully."
 read -p "Press enter when you have tried to save your work and ask it nicely to close."
 
 killall -9 Godot 2> /dev/null || true
 sleep 2
-osascript -e 'tell application "Terminal" to activate'
 
 # Find all .bsp files and store them in an array
 bsp_files=($(find . -name "*.bsp"))
@@ -14,19 +23,13 @@ bsp_files=($(find . -name "*.bsp"))
 echo "Deleting all bsp files"
 rm assets/maps/*.bsp* 2> /dev/null || true
 
-open project.godot
-sleep 3
-osascript -e 'tell application "Terminal" to activate'
-#read -p "Now start GoDot and open the project, then return here and press enter to continue."
-
 # Loop through the same files and run git checkout to restore them
 for file in "${bsp_files[@]}"; do
-    echo "Restoring $file"
+    echo "Restoring and importing $file"
     git checkout -- "$file"
-    osascript -e 'tell application "Godot" to activate'
-    sleep 3
-    osascript -e 'tell application "Terminal" to activate'
-    # read -p "Switch to GoDot and wait for the import to finish, then come back here and press enter."
+    "$godot" --headless --import
 done
+
+open project.godot
 
 echo "Done!"
